proyecto_musica/requirements.txt
django==3.2.4
djangorestframework==3.13.1
PyMySQL==1.0.2
PyJWT==2.4.0
daphne==4.0.0
channels==4.0.0
channels-redis
Pillow==9.2.0
channels-redis==4.0.0
Django==3.2.4
django-cors-headers==3.13.0
djangorestframework-simplejwt==5.2.0
mysqlclient==2.1.1
redis==4.3.4
requests==2.28.2
